# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        example: Credenciales inválidas
        type: string
    type: object
...
  controllers.LoginData:
    properties:
      email:
        example: user@example.com
        type: string
      module_key:
        example: '.....'
        type: string
      password:
        example: securePassword123
        type: string
    type: object
...
  controllers.TokenResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  models.Module:
    properties:
      active:
        type: boolean
      created_at:
        description: |-
          CreatedAt   time.Time `gorm:"default:CURRENT_TIMESTAMP" json:"created_at"`
          UpdatedAt   time.Time `gorm:"default:CURRENT_TIMESTAMP;autoUpdateTime" json:"updated_at"`
        type: string
      description:
        type: string
      id:
        type: integer
      module_key:
        type: string
      name:
        type: string
      permissions:
        description: Permissions []Permission `gorm:"foreignKey:ModuleID" json:"permissions"`
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.Permission:
    properties:
      active:
        type: boolean
      created_at:
        description: |-
          CreatedAt   time.Time `gorm:"default:CURRENT_TIMESTAMP" json:"created_at"`
          UpdatedAt   time.Time `gorm:"default:CURRENT_TIMESTAMP;autoUpdateTime" json:"updated_at"`
        type: string
      description:
        type: string
      id:
        type: integer
      module:
        $ref: '#/definitions/models.Module'
      module_id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: seri-api-utn-2024.fly.dev
info:
  contact:
    email: jfpastillod@utn.edu.ec
    name: Pastillo D Joan
    url: https://www.utn.edu.ec
  description: Documentación de LA API DE SEGURIDAD hecha con Go.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: API SEGURIDAD
  version: "1.9"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Este endpoint autentica un usuario utilizando su email, contraseña
        y la clave del módulo correspondiente. Si la autenticación es exitosa, se
        genera y devuelve un token JWT.
      parameters:
      - description: Datos de inicio de sesión (email, password y la key del módulo
          correspondiente)
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginData'
      produces:
      - application/json
      responses:
        "200":
          description: Token JWT generado
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
        "400":
          description: Datos inválidos en la solicitud
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Credenciales inválidas
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Error interno al procesar la autenticación
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Iniciar sesión
      tags:
      - Autenticación
  /logout:
    post:
      description: Invalida la sesión actual del usuario. Requiere un Bearer Token
        válido. Este endpoint cierra la sesión del usuario, eliminando el acceso al
        sistema hasta una nueva autenticación.
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cerrar sesión
      tags:
      - Autenticación
...
  /permissions:
    get:
      description: Devuelve una lista de permisos paginada.
      parameters:
      - description: 'Número de página (por defecto: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamaño de página (por defecto: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: permissions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener permisos
      tags:
      - Permisos
    post:
      consumes:
      - application/json
      description: Crea un nuevo permiso con nombre, descripción, ID del módulo y
        estado activo. Requiere un Bearer Token.
      parameters:
      - description: Datos del permiso
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: permission
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Crear permiso
      tags:
      - Permisos
  /permissions/{id}:
    get:
      description: Devuelve los detalles de un permiso dado su ID. Requiere un Bearer
        Token.
      parameters:
      - description: ID del permiso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: permission
          schema:
            $ref: '#/definitions/models.Permission'
        "401":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obtener permiso por ID
      tags:
      - Permisos
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un permiso existente. Requiere un Bearer
        Token.
      parameters:
      - description: ID del permiso
        in: path
        name: id
        required: true
        type: string
      - description: Datos del permiso
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses:
        "200":
          description: permission
          schema:
            $ref: '#/definitions/models.Permission'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar permiso
      tags:
      - Permisos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"